% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nn_loglike}
\alias{nn_loglike}
\title{Neural Network Normal Log-likelihood Value}
\usage{
nn_loglike(nn_model)
}
\arguments{
\item{nn_model}{nnet object}
}
\value{
Log-Likelihhod value
}
\description{
Neural Network Normal Log-likelihood Value
}
