% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_functions.R, R/utils.R
\name{nn_pred}
\alias{nn_pred}
\title{Neural network prediction}
\usage{
nn_pred(X, W, q)

nn_pred(X, W, q)
}
\arguments{
\item{X}{Data}

\item{W}{Weight vector}

\item{q}{Number of hidden units}

\item{output}{Activation function for output unit: \code{"identity"} (default) or
\code{"sigmoid"}}
}
\value{
Prediction for given inputs

Prediction for given inputs
}
\description{
Neural network prediction

Network Output Function
}
