% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statnnet-package.R
\docType{package}
\name{statnnet-package}
\alias{statnnet-package}
\title{statnnet: statistical summaries for fitted nnet models}
\description{
\code{statnnet} provides regression-style statistical summaries for small,
parsimonious, single-hidden-layer neural networks already fitted with
\code{\link[nnet:nnet]{nnet::nnet()}}. It does not refit, select, or provide post-selection
inference for neural networks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andrew-mcinerney/statnnet}
  \item \url{https://andrew-mcinerney.github.io/statnnet/}
  \item Report bugs at \url{https://github.com/andrew-mcinerney/statnnet/issues}
}

}
\author{
\strong{Maintainer}: Andrew McInerney \email{andrew.mcinerney@ul.ie} (\href{https://orcid.org/0000-0002-2348-2159}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Andrew McInerney \email{andrew.mcinerney@ul.ie} (\href{https://orcid.org/0000-0002-2348-2159}{ORCID}) [copyright holder]
  \item Kevin Burke \email{kevin.burke@ul.ie}
}

}
\keyword{internal}
