% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statnnet.R
\name{statnnet}
\alias{statnnet}
\title{statnnet}
\usage{
statnnet(nn, X, B = 1000)
}
\arguments{
\item{nn}{nnet object}

\item{X}{matrix of values for nnet}

\item{B}{number of bootstrap replicates}
}
\value{
statnnet object
}
\description{
statnnet
}
